refactor: extract data-loaders.js
